New feature: facilitate summarisation of SyncSeqApply.map results.
A common scenario is for the sequence generated by map to be summarised by simply yielding the number of database rows affected by (the various executions of) the apply function versus the total number of executions of the apply function. The new summarize function provides that simple mechanism, along with a facility to produce a side effect for each execution of apply (e.g. to log success/failure).